requirements.txt
# MCP Server
mcp>=0.9.0

# Database
supabase>=2.0.0
psycopg2-binary>=2.9.9
pgvector>=0.2.0

# AI/ML
openai>=1.0.0
anthropic>=0.25.0
sentence-transformers>=2.0.0
google-generativeai>=0.3.0

# Async support
asyncio
aiohttp>=3.9.0
aiofiles>=22.1.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
holidays>=0.21.0

# REST API
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Document Processing
PyPDF2>=3.0.0
python-docx>=0.8.11

# Testing (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0
